+++ src/SmtpServer/ComponentModel/ServiceProvider.cs
@@ -19,6 +19,7 @@ namespace SmtpServer.ComponentModel
         IEndpointListenerFactory _endpointListenerFactory;
         IUserAuthenticatorFactory _userAuthenticatorFactory;
         ISmtpCommandFactory _smtpCommandFactory;
+        ISmtpCommandPolicyFactory _smtpCommandPolicyFactory;
         IMailboxFilterFactory _mailboxFilterFactory;
         IMessageStoreFactory _messageStoreFactory;
 
@@ -30,6 +31,7 @@ namespace SmtpServer.ComponentModel
             Add(UserAuthenticator.Default);
             Add(MailboxFilter.Default);
             Add(MessageStore.Default);
+            Add(SmtpCommandPolicy.Default);
         }
 
         /// <summary>
@@ -66,6 +68,24 @@ namespace SmtpServer.ComponentModel
         public void Add(ISmtpCommandFactory smtpCommandFactory)
         {
             _smtpCommandFactory = smtpCommandFactory;
+        }
+
+        /// <summary>
+        /// Add an instance of the SMTP command policy factory.
+        /// </summary>
+        /// <param name="smtpCommandPolicyFactory">The SMTP command policy factory.</param>
+        public void Add(ISmtpCommandPolicyFactory smtpCommandPolicyFactory)
+        {
+            _smtpCommandPolicyFactory = smtpCommandPolicyFactory;
+        }
+
+        /// <summary>
+        /// Add an instance of the SMTP command policy.
+        /// </summary>
+        /// <param name="smtpCommandPolicy">The SMTP command policy.</param>
+        public void Add(ISmtpCommandPolicy smtpCommandPolicy)
+        {
+            _smtpCommandPolicyFactory = new DelegatingSmtpCommandPolicyFactory(context => smtpCommandPolicy);
         }
 
         /// <summary>
@@ -124,6 +144,11 @@ namespace SmtpServer.ComponentModel
             if (serviceType == typeof(ISmtpCommandFactory))
             {
                 return _smtpCommandFactory;
+            }
+
+            if (serviceType == typeof(ISmtpCommandPolicyFactory))
+            {
+                return _smtpCommandPolicyFactory;
             }
 
             if (serviceType == typeof(IMailboxFilterFactory))