+++ src/SmtpServer/Tracing/TracingSmtpCommandVisitor.cs
@@ -75,6 +75,33 @@ namespace SmtpServer.Tracing
         }
 
         /// <summary>
+        /// Visit a HELP command.
+        /// </summary>
+        /// <param name="command">The command that is being visited.</param>
+        protected override void Visit(HelpCommand command)
+        {
+            _output.WriteLine("HELP: Argument={0}", command.Argument);
+        }
+
+        /// <summary>
+        /// Visit a VRFY command.
+        /// </summary>
+        /// <param name="command">The command that is being visited.</param>
+        protected override void Visit(VrfyCommand command)
+        {
+            _output.WriteLine("VRFY: Argument={0}", command.Argument);
+        }
+
+        /// <summary>
+        /// Visit an EXPN command.
+        /// </summary>
+        /// <param name="command">The command that is being visited.</param>
+        protected override void Visit(ExpnCommand command)
+        {
+            _output.WriteLine("EXPN: Argument={0}", command.Argument);
+        }
+
+        /// <summary>
         /// Visit an NOOP command.
         /// </summary>
         /// <param name="command">The command that is being visited.</param>